requirements.txt
alembic==1.14.0
altgraph==0.17.4
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
codecov==2.1.13
coverage==7.6.10
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
freezegun==1.5.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
macholib==1.16.3
Mako==1.3.8
MarkupSafe==3.0.2
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.6
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
ruff==0.8.4
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
